# openapi.yml
openapi: 3.1.0
info:
  title: Система бронирования мест в кафе
  version: 0.0.2
servers:
  - url: http://localhost:8000/
    description: Локальный сервер
...
paths:
...
  /cafe/{cafe_id}/tables:
    get:
      tags:
        - Столы
      summary: Список столов в кафе
      description: Получение списка доступных для бронирования столов в кафе. Для администраторов и менеджеров - все столы (с возможностью выбора), для пользователей - только активные.
      operationId: get_all_tables_cafe__cafe_id__tables_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
        - name: show_all
          in: query
          required: false
          schema:
            type: boolean
            title: Показывать все столы?
            description: Показывать все столы в кафе или нет. По умолчанию показывает все столы
            default: false
          description: Показывать все столы в кафе или нет. По умолчанию показывает все столы
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableInfo'
                title: Response Get All Tables Cafe  Cafe Id  Tables Get
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
    post:
      tags:
        - Столы
      summary: Новый стол в кафе
      description: Создает новое стола кафе. Только для администраторов и менеджеров.
      operationId: create_table_cafe__cafe_id__tables_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableCreate'
              description: Данные нового стола
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
  /cafe/{cafe_id}/tables/{table_id}:
    get:
      tags:
        - Столы
      summary: Информация о столе в кафе по его ID
      description: Получение информации о столе в кафе по его ID. Для администраторов и менеджеров - все столы, для пользователей - только активные.
      operationId: get_table_by_id_cafe__cafe_id__tables__table_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
        - name: table_id
          in: path
          required: true
          schema:
            type: integer
            description: ID стола
            title: Table Id
          description: ID стола
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
    patch:
      tags:
        - Столы
      summary: Обновление информации о столе в кафе по его ID
      description: Обновление информации о столе в кафе по его ID. Только для администраторов и менеджеров.
      operationId: path_table_by_id_cafe__cafe_id__tables__table_id__patch
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
        - name: table_id
          in: path
          required: true
          schema:
            type: integer
            description: ID стола
            title: Table Id
          description: ID стола
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableUpdate'
              description: Обновленные данные стола
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
...
  /media/{media_id}:
    get:
      tags:
        - Медиа
      summary: Возвращает изображение в бинарном формате
      operationId: get_media_media__media_id__get
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: ID изображения
            title: Media Id
          description: ID изображения
      responses:
        '200':
          description: Успешно. Возвращает изображение в бинарном формате
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
  /media:
    post:
      tags:
        - Медиа
      summary: Загрузка изображения
      description: Загрузка изображения на сервер. Поддерживаются форматы jpg, png. Размер файла не более 5Мб. Только для администраторов и менеджеров
      operationId: upload_media_media_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MediaData'
        required: true
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка сохранения файла
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
      security:
        - OAuth2PasswordBearer: []
components:
  schemas:
...
    MediaData:
      properties:
        file:
          type: string
          format: binary
          title: File
          description: Загружаемый файл
      type: object
      required:
        - file
      title: MediaData
...
    CafeShortInfo:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        address:
          type: string
          title: Address
        phone:
          type: string
          title: Phone
        description:
          type: string
          title: Description
        photo_id:
          type: string
          format: uuid
          title: Photo Id
      type: object
      required:
        - id
        - name
        - address
        - phone
        - description
        - photo_id
      title: CafeShortInfo
...
    CustomError:
      properties:
        code:
          type: integer
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
        - code
        - message
      title: CustomError
...
    MediaInfo:
      properties:
        media_id:
          type: string
          format: uuid
          title: Media Id
      type: object
      required:
        - media_id
      title: MediaInfo
    TableCreate:
      properties:
        cafe_id:
          type: integer
          title: Cafe Id
        description:
          type: string
          title: Description
        seat_number:
          type: integer
          title: Seat Number
      type: object
      required:
        - cafe_id
        - description
        - seat_number
      title: TableCreate
    TableInfo:
      properties:
        id:
          type: integer
          title: Id
        cafe:
          $ref: '#/components/schemas/CafeShortInfo'
        description:
          type: string
          title: Description
        seat_number:
          type: integer
          title: Seat Number
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
        - id
        - cafe
        - description
        - seat_number
        - is_active
        - created_at
        - updated_at
      title: TableInfo
...
    TableUpdate:
      properties:
        cafe_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Cafe Id
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        seat_number:
          anyOf:
            - type: integer
            - type: 'null'
          title: Seat Number
        is_active:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Active
      type: object
      required:
        - cafe_id
        - description
        - seat_number
        - is_active
      title: TableUpdate
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login